requirements.txt
Django~=4.0.1
djangorestframework~=3.13.1
requests~=2.27.1
